packages/kindling_otel_azure/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "kindling-otel-azure"
version = "0.1.0.4"
description = "Azure Monitor OpenTelemetry integration for Kindling framework"
authors = ["SEP Engineering <engineering@sep.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "kindling_otel_azure"},
]

[tool.poetry.dependencies]
python = "^3.10"
# kindling is provided by the runtime environment, not installed as a dependency
# Pin to exact compatible versions verified working in Fabric (LogData available)
azure-monitor-opentelemetry = "1.2.0"
opentelemetry-api = "1.21.0"
opentelemetry-sdk = "1.21.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
black = ">=23.0.0"
